algo: support 'Sensor Delay Measure'
Run self measurement of the sensor delay by setting different
values of sensor gain and exposure and measuring the number of
VSYNC it takes to get the statistic updated.

The implementation relies on an algorithm started upon an
incoming GET SENSOR DELAY MEASURE command issued by the remote
IQTune. When all the measures have been done, which can take
up to 4 seconds, the algorithm sends the measured delay in a
GET answer message to the remote.

The algorithm (available only when ISP_MW_TUNING_TOOL_SUPPORT
is defined) suspends the other algorithms (AEC, AWB) to avoid
interference, and restore them at the end of the measures.

Signed-off-by: Fabien Dessenne <fabien.dessenne@st.com>