Support Restart State mode
This mode allows the ISP to start with last ISP state, i.e. with the
latest sensor/ISP settings applied by the AEC / AWB algorithms.
As a typical use case, when the system wakes up from standby, the ISP
immediately displays a correct frame instead of a black frame.

ISP continuously updates the AEC/AWB config in the ISP_RestartState
variable and upon restart it applies this config so the first output
frames are correct.

The feature is enabled when the application calls ISP_EnableRestartState()
providing a pointer where ISP can store ISP_RestartState.
To fully support the feature in a Low Power use Case ISP_RestartState
must be allocated in some "Retention RAM" memory.

Signed-off-by: Fabien Dessenne <fabien.dessenne@st.com>